Codes addition

+++ src/main/kotlin/org/l11_3/session/JWTController.kt
@@ -8,11 +8,11 @@ private const val userID = "userID"
 private const val password = "password"
 private const val ONE_YEAR_IN_MS = 31_536_000_000
 
-fun generateToken(values: JWTValues, id: Int, userPassword: Int): String = JWT
+fun generateToken(values: JWTValues, id: UInt, userPassword: Int): String = JWT
     .create()
     .withAudience(values.audience)
     .withIssuer(values.issuer)
-    .withClaim(userID, id)
+    .withClaim(userID, id.toInt())
     .withClaim(password, userPassword)
     .withExpiresAt(Date(System.currentTimeMillis() + ONE_YEAR_IN_MS))
     .sign(getAlgorithm(values.secret))
